🛰️ `/[short]` test
